tree: refactor defaults folder structure (#18332)
## Description

This PR improves the clarity of the defaults folder and optional field
by moving optional field (and related codecs/types) out of the defaults
folder and into an optional-field folder (a sibling of sequence-field).
It also renames the default folder to default-schema to match
modular-schema.
