+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -214,7 +214,7 @@ export {
 	OptionalFieldEditBuilder,
 	SequenceFieldEditBuilder,
 	defaultSchemaPolicy,
-} from "./default-field-kinds";
+} from "./default-schema";
 
 export {
 	AssignableFieldKinds,
