+++ experimental/dds/tree2/src/feature-libraries/optional-field/README.md
@@ -1,0 +1,3 @@
+# optional-field
+
+A `FieldChangeHandler` for optional fields for use in [`modular-schema`](../../modular-schema/README.md).
