+++ experimental/dds/tree2/src/feature-libraries/optional-field/index.ts
@@ -1,0 +1,13 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export { noChangeCodecFamily, makeOptionalFieldCodecFamily } from "./optionalFieldCodecs";
+export { OptionalChangeset } from "./optionalFieldChangeTypes";
+export {
+	optionalChangeHandler,
+	optionalFieldEditor,
+	optionalChangeRebaser,
+	optionalFieldIntoDelta,
+} from "./optionalField";
