+++ experimental/dds/tree2/src/feature-libraries/sequence-field/README.md
@@ -1,6 +1,6 @@
 # sequence-field
 
-A sequence-shaped `FieldChangeHandler` for use in [`modula-schema`](../../modula-schema/README.md).
+A sequence-shaped `FieldChangeHandler` for use in [`modular-schema`](../../modular-schema/README.md).
 
 This implementation is incomplete.
 See comments on rebase, compose, and invert functions.
