+++ experimental/dds/tree2/src/test/feature-libraries/default-field-kinds/defaultFieldKinds.spec.ts
@@ -11,7 +11,7 @@ import {
 	valueFieldEditor,
 	// Allow import from file being tested.
 	// eslint-disable-next-line import/no-internal-modules
-} from "../../../feature-libraries/default-field-kinds/defaultFieldKinds";
+} from "../../../feature-libraries/default-schema/defaultFieldKinds";
 import {
 	makeAnonChange,
 	TaggedChange,
@@ -22,8 +22,8 @@ import {
 import { brand, fakeIdAllocator } from "../../../util";
 import { defaultRevisionMetadataFromChanges } from "../../utils";
 // eslint-disable-next-line import/no-internal-modules
-import { OptionalChangeset } from "../../../feature-libraries/default-field-kinds/defaultFieldChangeTypes";
-import { changesetForChild, testTree, testTreeCursor } from "./fieldKindTestUtils";
+import { OptionalChangeset } from "../../../feature-libraries/optional-field";
+import { changesetForChild, testTree, testTreeCursor } from "../fieldKindTestUtils";
 
 /**
  * A change to a child encoding as a simple placeholder string.
