+++ experimental/dds/tree2/src/test/feature-libraries/fieldKindTestUtils.ts
@@ -3,10 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import { FieldKinds, NodeChangeset, cursorForJsonableTreeNode } from "../../../feature-libraries";
-import { FieldKey, JsonableTree, ITreeCursorSynchronous } from "../../../core";
-import { brand } from "../../../util";
-import { leaf } from "../../../domains";
+import { FieldKinds, NodeChangeset, cursorForJsonableTreeNode } from "../../feature-libraries";
+import { FieldKey, JsonableTree, ITreeCursorSynchronous } from "../../core";
+import { brand } from "../../util";
+import { leaf } from "../../domains";
 
 // TODO: Users of this are mainly working with in memory representations.
 // Therefore it should not be using JsonableTrees.
